duplicate directory from main branch, seperated
