--- flask_app/birdsfinal/flask_app/static/style.css
@@ -1,9 +1,0 @@
-#background-image{
-    background-image: url('./img/Tropical\ birds-amico.png'); 
-    height: 100vh;
-    width:100vw;
-}
-
-#navie{
-    width:100vw; background-color: #D8E0F5; opacity: .8;
-}
