--- flask_app/birdsfinal/flask_app/templates/index.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-    <link rel="stylesheet" href="../static/style.css">
-    <title>Index Login</title>
-</head>
-
-<body class="">
-    <div id="background-image">
-        <!-- navigation -->
-        <div id="navie" class="navbar p-3">
-            <a class="navbar-brand mx-2"  href="/dashboard">Rate My Birds</a>
-            <div class="navbar-item">
-                <div class="navbar-text me-2"><a href="/logout">Logout</a></div>
-            </div>
-        </div>
-        <div class="row" id="mainBody">
-            <div class="col-5 container">
-                <div class="card-body m-lg-5">
-                    <h2 class="m-2">Birdly News</h2>
-                    <!-- https://rapidapi.com/newscatcher-api-newscatcher-api-default/api/newscatcher/ -->
-                    <p class="card-text"> Sample Text Sample Text Sample Text Sample Text Sample Text Sample Text Sample
-                        Text Sample Text Sample Text
-                        Sample Text Sample Text Sample Text Sample Text Sample Text Sample Text Sample Text
-                        Sample Text Sample Text Sample Text Sample Text Sample Text Sample Text
-                    </p>
-                </div>
-            </div>
-            <div class="col-sm-7 col-md-7 col-lg-5 col-sm-auto m-5 card-body container-fluid bg-opacity-10 bg-danger"
-                style="border-radius: 5px;">
-                <h2 class="m-2">Join Birds</h2>
-                <!-- Show Flash Errors -->
-                {% with messages = get_flashed_messages() %}
-                {% if messages %}
-                {% for message in messages %}
-                <p>{{message}}</p>
-                {% endfor %}
-                {% endif %}
-                {% endwith %}
-                <div class="m-2">
-                    <!-- Register Div -->
-                    <div class="card-body">
-                        <h4 class="card-header">Register</h4>
-                        <!-- Reg Form -->
-                        <form action="/register" method="post">
-                            <div class="form-group">
-                                <label for="firstname">First Name</label>
-                                <input class="form-control" type="text" name="firstname">
-                            </div>
-                            <div class="form-group">
-                                <label for="lastname">Last Name</label>
-                                <input class="form-control" type="text" name="lastname">
-                            </div>
-                            <div class="form-group">
-                                <label for="email">Email: </label>
-                                <input class="form-control" type="text" name="email">
-                            </div>
-                            <div class="form-group">
-                                <label for="password">Password: </label>
-                                <input class="form-control" type="text" name="password">
-                            </div>
-                            <div class="form-group">
-                                <label for="confirm">Confirm Password: </label>
-                                <input class="form-control" type="text" name="confirm">
-                            </div>
-                            <div>
-                                <button class="btn btn-primary mt-1" type="submit">Submit</button>
-                            </div>
-                        </form>
-                    </div>
-
-                    <!-- Login -->
-                    <div class="card-body">
-                        <h4 class="card-header">Login</h4>
-                        <!-- Login Form -->
-                        <form action="/login" method="post">
-                            <div class="form-group">
-                                <label for="email">Email: </label>
-                                <input class="form-control" type="text" name="email">
-                            </div>
-                            <div class="form-group">
-                                <label for="password">Password: </label>
-                                <input class="form-control" type="text" name="password">
-                            </div>
-                            <div>
-                                <button class="btn btn-primary mt-1" type="submit">Submit</button>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="modal-footer"><a href="https://storyset.com/nature">Nature illustrations by Storyset</a></div>
-    </div>
-</body>
-
-</html>
