+++ flask_app/models/user.py
@@ -14,6 +14,7 @@ class User:
         self.password = data['password']
         self.created_at = data['created_at']
         self.updated_at = data['updated_at']
+        # self.birds = []
         # birds=[]
         # ratings=[]
