+++ flask_app/static/style.css
@@ -1,5 +1,5 @@
 #background-image{
-    background-image: url('./img/tropicalbirdsamico.png'); 
+    background-image: url('./img/Tropical\ birds-amico.png'); 
     height: 100vh;
     width:100vw;
 }
