--- flask_app/templates/oldloginreg.html
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-    <link rel="stylesheet" href="../static/style.css">
-    <title>LoginReg</title>
-</head>
-<body class="bg-light">
-        <!-- navigation -->
-    <div id="navie" class="navbar p-3">
-        <a class="navbar-brand mx-2">Rate My Birds</a>
-        <div class="navbar-item">
-            <div class="navbar-text me-2">Sample Text</div>
-        </div>
-    </div>
-    <!-- Show Flash Errors -->
-    {% with messages = get_flashed_messages() %}
-        {% if messages %}
-            {% for message in messages %}
-                <p>{{message}}</p>
-            {% endfor %}
-        {% endif %}
-    {% endwith %}
-    
-    <!-- main div -->
-    <div class="row align-content-center mr-auto">
-
-        <!-- Register -->
-        <div class="card-body">
-            <h4 class="card-header">Register</h4>
-            <form action="/create/user" method="post">
-                <div class="form-group">
-                    <label for="fname">First Name</label>
-                    <input class="form-control" type="text" name="fname">
-                </div>
-                <div class="form-group">
-                    <label for="lname">Last Name</label>
-                    <input class="form-control" type="text" name="lname">
-                </div>
-                <div class="form-group">
-                    <label for="email">Email: </label>
-                    <input class="form-control" type="text" name="email">
-                </div>
-                <div class="form-group">
-                    <label for="password">Password: </label>
-                    <input class="form-control" type="text" name="password">
-                </div>
-                <div class="form-group">
-                    <label for="confirm">Confirm Password: </label>
-                    <input class="form-control" type="text" name="confirm">
-                </div>
-                <div>
-                    <button class="btn btn-primary" type="submit">Submit</button>
-                </div>
-            </form>
-        </div>
-
-        <!-- Login -->
-        <div class="card-body">
-            <h4 class="card-header">Login</h4>
-            <form action="/login" method="post">
-                <div class="form-group">
-                    <label for="email">Email: </label>
-                    <input class="form-control" type="text" name="email">
-                </div>
-                <div class="form-group">
-                    <label for="password">Password: </label>
-                    <input class="form-control" type="text" name="password">
-                </div>
-                <div>
-                    <button class="btn btn-primary" type="submit">Submit</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</body>
-</html>
