+++ flask_app/templates/oneBird.html
@@ -8,15 +8,14 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
     <link rel="stylesheet" href="../static/style.css">
-
-    <title>Document</title>
+    <title>View Bird</title>
 </head>
 
 <body>
     <div id="background-image">
         <!-- navigation -->
         <div id="navie" class="navbar p-3">
-            <a class="navbar-brand mx-2" href="/dashboard">Rate My Birds</a>
+            <a class="navbar-brand mx-2"  href="/dashboard">Rate My Birds</a>
             <div class="navbar-item">
                 <div class="navbar-text me-2"><a href="/logout">Logout</a></div>
             </div>
@@ -45,7 +44,7 @@
                     <div class="container">
                         <span id="rateMe1"></span>
                     </div>
-                    <form class="form">
+                    <form class="">
                         <label for="rating_value"></label>
                         <select name="rating_value" id="">
                             <option value="1">1: Weak bird</option>
@@ -63,5 +62,6 @@
     </div>
 
 </body>
+<script src="../static/js/rating.js"></script>
 
 </html>
