+++ flask_app/templates/profile.html
@@ -7,54 +7,55 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-    <link rel="stylesheet" href="../static/style.css">
+        <link rel="stylesheet" href="../static/style.css">
     <title>Profile</title>
 </head>
 
 <body>
     <div id="background-image">
-        <!-- navigation -->
-        <div id="navie" class="navbar p-3">
-            <a class="navbar-brand mx-2" href="/dashboard">Rate My Birds</a>
-            <div class="navbar-item">
-                <div class="navbar-text me-2"><a href="/logout">Logout</a></div>
+    <!-- navigation -->
+    <div id="navie" class="navbar p-3">
+        <a class="navbar-brand mx-2"  href="/dashboard">Rate My Birds</a>
+        <div class="navbar-item">
+            <div class="navbar-text me-2"><a href="/logout">Logout</a></div>
+        </div>
+    </div>
+    <!-- body -->
+    <div class="main row justify-content-around" id="mainBody">
+        <div id="page-title" class="h3" style="text-align: center; min-width: 500px">Welcome to My Page</div>
+        <!-- left side -->
+        <div class="col row align-content-start"  style="margin: 20px; min-width: 550px; max-width: 650px;">
+            <div id="left" class="col card bg-opacity-10 bg-danger">
+                <h4></h4>
+                <!-- Picture, Information -->
+                <div class="bg-image hover-zoom" style="min-width: 500px; max-width: 600px; margin: auto;">
+                    <img src="../static/img/rust_sky.png" alt="test img" class="img-thumbnail">
+                </div>
+                Name: {{user.firstname}} <br/>
+                Favorite Bird: Tweety
             </div>
         </div>
-        <!-- body -->
-        <div class="main row justify-content-around" id="mainBody">
-            <div id="page-title" class="h3" style="text-align: center; min-width: 500px">Welcome to My Page</div>
-            <!-- left side -->
-            <div class="col row align-content-start" style="margin: 20px; min-width: 550px; max-width: 650px;">
-                <div id="left" class="col card bg-opacity-10 bg-danger">
-                    <h4></h4>
-                    <!-- Picture, Information -->
-                    <div class="bg-image hover-zoom" style="min-width: 500px; max-width: 600px; margin: auto;">
-                        <img src="../static/img/rust_sky.png" alt="test img" class="img-thumbnail">
-                    </div>
-                    Name: {{user.firstname}} <br />
-                    Favorite Bird: Tweety
-                </div>
-            </div>
-            <!-- right side -->
-            <div class="col row align-content-start" style="margin: 20px; min-width: 300px; max-width: 650px;">
-                <div class="col card-body container-fluid">
-                    <h4>Created Birds</h4>
-                    <ul class="list-group">
-                        {% for bird in birds %}
-                        <li class="list-group-item">{{bird.species}}</li>
-                        {% endfor %}
-                        <li class="list-group-item">Cardinal(-Test)</li>
-                        <li class="list-group-item">Tweety(-Test)</li>
-                        <li class="list-group-item">Emu(-Test)</li>
-                        <li class="list-group-item">Falcon(-Test)</li>
-                        <li class="list-group-item">Pelican(-Test)</li>
-                    </ul>
-                </div>
+        <!-- right side -->
+        <div class="col row align-content-start"  style="margin: 20px; min-width: 300px; max-width: 650px;">
+            <div class="col card-body container-fluid">
+                <h4>Created Birds</h4>
+                <ul class="list-group">
+                    {% for bird in birds %}
+                <li class="list-group-item">{{bird.species}}</li>
+                    {% endfor %}
+                    <li class="list-group-item">Cardinal(-Test)</li>
+                    <li class="list-group-item">Tweety(-Test)</li>
+                    <li class="list-group-item">Emu(-Test)</li>
+                    <li class="list-group-item">Falcon(-Test)</li>
+                    <li class="list-group-item">Pelican(-Test)</li>
+                </ul>
             </div>
         </div>
-
     </div>
 
+    </div>
 </body>
+
+</html>
 
 </html>
